aspsp-mock-server/src/main/java/de/adorsys/aspsp/aspspmockserver/service/PaymentService.java
package de.adorsys.aspsp.aspspmockserver.service;

import de.adorsys.aspsp.aspspmockserver.repository.PaymentRepository;
import de.adorsys.aspsp.xs2a.spi.domain.payment.SpiPaymentInitialisationResponse;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class PaymentService {
    private PaymentRepository paymentRepository;

    @Autowired
    public PaymentService(PaymentRepository paymentRepository) {
        this.paymentRepository = paymentRepository;
    }

    public SpiPaymentInitialisationResponse addPayment(SpiPaymentInitialisationResponse payment) {
        return paymentRepository.save(payment);
    }
}
